[fix] Line layer is not displayed for between hex ids
