+++ src/layers/src/arc-layer/arc-layer.ts
@@ -14,9 +14,8 @@ import {BrushingExtension} from '@deck.gl/extensions';
 import {GeoArrowArcLayer} from '@kepler.gl/deckgl-arrow-layers';
 import {FilterArrowExtension} from '@kepler.gl/deckgl-layers';
 import {ArcLayer as DeckArcLayer} from '@deck.gl/layers';
-import {h3ToGeo} from 'h3-js';
 
-import {hexToRgb, DataContainerInterface, ArrowDataContainer} from '@kepler.gl/utils';
+import {hexToRgb, maybeHexToGeo, DataContainerInterface, ArrowDataContainer} from '@kepler.gl/utils';
 import ArcLayerIcon from './arc-layer-icon';
 import {isLayerHoveredFromArrow, createGeoArrowPointVector, getFilteredIndex} from '../layer-utils';
 import {
@@ -138,33 +137,6 @@ const DEFAULT_COLUMN_MODE = COLUMN_MODE_POINTS;
 
 const brushingExtension = new BrushingExtension();
 const arrowCPUFilterExtension = new FilterArrowExtension();
-
-export function getPositionFromHexValue(token) {
-  const pos = h3ToGeo(token);
-
-  if (Array.isArray(pos) && pos.every(Number.isFinite)) {
-    return [pos[1], pos[0]];
-  }
-  return null;
-}
-
-function maybeHexToGeo(
-  dc: DataContainerInterface,
-  d: {index: number},
-  lat: LayerColumn,
-  lng: LayerColumn
-) {
-  // lat or lng column could be hex column
-  // we assume string value is hex and try to convert it to geo lat lng
-  const latVal = dc.valueAt(d.index, lat.fieldIdx);
-  const lngVal = dc.valueAt(d.index, lng.fieldIdx);
-
-  return typeof latVal === 'string'
-    ? getPositionFromHexValue(latVal)
-    : typeof lngVal === 'string'
-    ? getPositionFromHexValue(lngVal)
-    : null;
-}
 
 function isOtherFieldString(columns, allFields, key) {
   const field = allFields[columns[key].fieldIdx];
